Factored EnsureSize interface out of common.py.
